refactor trie interface
It didn't feel right to have phone_words as a method on the Trie, but
moving it out to its own function would just be an equivalent signature
like phone_words(&Trie, &str). Then I realized that the whole point of
this is to be the main function, so I can just inline the search logic
and call it a day.

This then put the pressure back on to expose the proper API from the
Trie struct. With the implementation of the search done, I found out
that only a couple methods were needed, and they clean up nicely.

I removed Trie::contains, since it was only used from the tests to make
sure I was making the trie properly, but was otherwise dead code. As
such, I also removed the tests and could finally remove the
#![allow(dead_code)].