inline consts as literals
These still apparently have a static lifetime, which works fine for me,
and the compiler can probably still figure out smart ways of allocating
this stuff. Haven't actually delved into the assembly output, but the
compiler is a big boy, so I reckon the difference is going to be
negligible at best. Versus the gains in readability & semantics from not
having these random constants lying around for what should really just
be literals.